+++ packages/world-module-erc20/test/ERC20Pausable.t.sol
@@ -5,6 +5,10 @@ import { Test } from "forge-std/Test.sol";
 import { console } from "forge-std/console.sol";
 
 import { GasReporter } from "@latticexyz/gas-report/src/GasReporter.sol";
+
+import { ResourceId } from "@latticexyz/store/src/ResourceId.sol";
+import { RESOURCE_TABLE } from "@latticexyz/store/src/storeResourceTypes.sol";
+
 import { IBaseWorld } from "@latticexyz/world/src/codegen/interfaces/IBaseWorld.sol";
 
 import { ERC20MetadataData } from "../src/codegen/tables/ERC20Metadata.sol";
@@ -12,10 +16,17 @@ import { IERC20Errors } from "../src/interfaces/IERC20Errors.sol";
 import { IERC20Events } from "../src/interfaces/IERC20Events.sol";
 import { MUDERC20 } from "../src/experimental/MUDERC20.sol";
 import { Pausable, ERC20Pausable } from "../src/experimental/ERC20Pausable.sol";
-import { MockERC20Base, ERC20BehaviorTest } from "./ERC20BaseTest.t.sol";
+import { MockERC20Base, ERC20BehaviorTest, namespace, pausedId } from "./ERC20BaseTest.t.sol";
 
 // Mock to include mint and burn functions
-contract MockERC20Pausable is MockERC20Base, ERC20Pausable {
+contract MockERC20Pausable is MockERC20Base, ERC20Pausable(pausedId) {
+  constructor(IBaseWorld world) MockERC20Base(world) {}
+
+  function initialize() public override {
+    super.initialize();
+    _Pausable_init();
+  }
+
   function pause() public {
     _pause();
   }
@@ -29,9 +40,9 @@ contract MockERC20Pausable is MockERC20Base, ERC20Pausable {
   }
 }
 
-abstract contract ERC20PausableBehaviorTest is ERC20BehaviorTest {
-  function createToken() internal override returns (MockERC20Base) {
-    return new MockERC20Pausable();
+contract ERC20PausableTest is ERC20BehaviorTest {
+  function createToken(IBaseWorld world) internal override returns (MockERC20Base) {
+    return new MockERC20Pausable(world);
   }
 
   function testPause() public {