+++ packages/react/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## 2.2.20
+
+### Patch Changes
+
+- Updated dependencies [06e48e0]
+  - @latticexyz/store@2.2.20
+  - @latticexyz/recs@2.2.20
+
 ## 2.2.19
 
 ### Patch Changes