+++ packages/recs/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## 2.2.20
+
+### Patch Changes
+
+- @latticexyz/schema-type@2.2.20
+- @latticexyz/utils@2.2.20
+
 ## 2.2.19
 
 ### Patch Changes