+++ packages/world-modules/CHANGELOG.md
@@ -1,5 +1,21 @@
 # Change Log
 
+## 2.2.20
+
+### Patch Changes
+
+- 3915759: Removed unsupported install methods as these now automatically revert in the base `Module` contract.
+- Updated dependencies [3187081]
+- Updated dependencies [06e48e0]
+- Updated dependencies [3915759]
+- Updated dependencies [06e48e0]
+- Updated dependencies [3187081]
+  - @latticexyz/world@2.2.20
+  - @latticexyz/store@2.2.20
+  - @latticexyz/common@2.2.20
+  - @latticexyz/config@2.2.20
+  - @latticexyz/schema-type@2.2.20
+
 ## 2.2.19
 
 ### Patch Changes